Create SpMVExample.c
I'll create a simple example that multiplies a sparse matrix with a dense vector using cuSPARSELt. This is a standard operation in many applications.

Unfortunately, as of my training cut-off in September 2021, cuSPARSELt doesn't have a high-level C++ API, only a C API. Therefore, the example code will be in C.

Here is a very basic example:

This example multiplies a sparse matrix stored in CSR format with a dense vector using the cuSPARSELt library. Note that the matrix and vector data are not filled in and you'll need to add this data yourself.

Please remember to link against the cuSPARSELt and CUDA libraries when compiling. You can save this program as SpMVExample.c.

Also note that, as of my training cut-off in September 2021, NVIDIA hasn't provided much detailed documentation or examples on how to use cuSPARSELt. The API may have changed or been expanded in the meantime. Always refer to the latest official documentation for the most accurate information.